refactor(tests): move the sheet-view property table into the test project
XLViewProperties.All has no production reader - the four consumers each list the properties in their own code, deliberately, so that adding one is a compile-time edit in three constructors rather than a reflection walk in the hottest file in the repo. That leaves a table of fifteen Action/Func pairs plus prose ("omitted when default; written (false) when non-default") shipping in the library, unreachable, and free to drift from the writer it describes with no build signal.

It is test data, so it now lives with the tests that read it. No behaviour change; the library binary simply stops carrying it.
